backend/requirements.txt
Flask==2.0.3
requests==2.26.0
gunicorn==20.1.0
Flask-Mail
Flask-Cors
python-dotenv
Werkzeug==2.0.3
Flask-PyMongo>=2.3.0
pymongo>=3.12
geopy
PyJWT